+++ src/components/resource/resourceTypeTable.js
@@ -182,9 +182,6 @@ export default function ResourceTable({ token, tokenData, env, resources, getThe
               indeterminate={numSelected > 0 && numSelected < rowCount}
               checked={rowCount > 0 && numSelected === rowCount}
               onChange={onSelectAllClick}
-              inputProps={{
-                'aria-label': 'select all desserts'
-              }}
             />
           </TableCell>
           {headCells.map((headCell) => (