+++ src/components/service/serviceChildren.js
@@ -172,9 +172,6 @@ export default function ServiceChildren({ masterTitle, setOpen, status, data, ge
               indeterminate={numSelected > 0 && numSelected < rowCount}
               checked={rowCount > 0 && numSelected === rowCount}
               onChange={onSelectAllClick}
-              inputProps={{
-                'aria-label': 'select all desserts'
-              }}
             />
           </TableCell>
           {headCells.map((headCell, index) => (